doc: fix curl example (#383)
Hello, it seems the curl example is not working without specifying the name of the header carrying the JWT. I lost a lot of time but learned a lot on Mercure. :) I don't know if it's the right process to suggest this change but I'm really late on my project, sorry! Cheers.